Add Basic Documentation
Added a section to the README file with some basic documentation for the widget. As mentioned in one of our meetings, I don't want to overload the README with too much detail. We should presume that someone using this theme knows WordPress enough to use widgets, and provide only enough detail to get them going. If more detailed information, step-by-step instructions, or multiple screenshots are required, we should link out to the Wiki