Fix Array Warnings in Main Menu
Menu items that were nested too deeply were causing PHP warnings about missing array keys. The line in our third processing loop that was presuming there was a 'parent' array in the parent of this menu item was the culprit. I added a check around that code to only attempt to insert these items when there was an appropriate destination for them.

I also added comments to code to help my understanding of what it was doing.