Add Visual Edit Button to Site Name in Footer
Added a visual shortcut button to the site name in the footer, per Zainab's note. This allows users to edit the site name from either location - they both point to the same setting/control in the Customizer.

To do this, I added a <span> inside the <h5> in the footer, and added that class to the list of classes targeted by the selective refresh settings in customizer-settings.php.