Remove Shortcut Button for Logo URL
Removed the selective refresh option for the alternate logo URL, which removed the visual shortcut button. There was not enough room in that general area for another button, and I'm thinking that one button for the logo makes more sense. Hopefully, someone wanting to edit the logo will click the current button and notice the two available options.