Change Transport Type of Contact Us Link
The customizer was refreshing the entire page when you modified the Contact Us link, when it should have been using the existing render function to draw just that part of the screen. Updated the 'transport' type to 'post Message' to tell WordPress to update just the section of the screen controlled by the render function.