Add Radio Buttons for Controlling Logo Type
Added new customizer controls and rendering logic for selecting a logo in the footer. The new standards call for the ASU logo to be used in the absence of any other logo having been chosen, so we now have two radio buttons for selecting 'ASU logo' or 'unit logo', essentially. If you choose 'unit logo', you can select from our drop-down, or provide a URL. I added fall-back logic so that neglecting to choose/provide a unit logo results in the ASU logo. Multiple changes here:

* add the actual radio button controls to customizer-settings.php
* update the logic in render-partials to match that stated above
* add and enqueue new javascript (customizer-controls-js.php) to show/hide the unit logo controls based on the radio button you select
* remove the 'ASU' logo from the endorsed logo drop-down, as it is now a separate choice