Update Code Checking Script to Ignore PHP < 5.6
We were running into some errors about the use of anonymous functions in some of our customizer code - despite the fact that the use of those anonymous functions is right in the example code for using 'selective refresh' - because it was not compatible with PHP 5.2 or earlier.

Since PHP 5.2 has been gone for a while, and our Travis CI settings are only going back as far as 5.6, I updated our 'check:cs' composer script to use a command line flag that only checks for issues pertinent to 5.6 or higher.