Fix Preview on InnerBlocks Example
Since our inner blocks sample doesn't have any content, I changed the preview mode to 'edit'. It's a bit confusing, but it means "when showing a little preview of this block, show what the ACF fields will look like, NOT the finished block."

Since this sample block doesn't have any content to preview (it's all going to be placed in the InnerBlock after you put this block on the page), there is nothing to actually show yet. Also, trying to show the preview in Chrome was actually causing a bug where you could not put the block on the page.