Remove Media Source and Image Fields
I had started down the path of allowing the user to choose between an image from the media library or a URL for the hero image. However, this is not implemented in the current template, so I'm removing these fields to avoid pushing them up to the repo until I can do everything at once.