Re-Key Card Block Group
Had to create a new key for this field group, as we had accidentally used an export of this group as the basis for the button block. This caused both the button and card field groups to have the same key, creating a conflict.

I generated a new key, using PHP's uniqid() method, and pasted it into the JSON export file I made, then re-imported the group to ACF.