Add CSS Fix for SVG Icons in Editor
The front-end SVG styles contained a `vertical-align: -.125rem` on all of the FontAwesome SVG icons. However, it was not working in the editor because the SVG's were all wrapped inside of some Gutenberg classes. I replicated the same rule, but specifically targeting SVG's inside of the Gutenberg classes, and enqueued that in the editor.