Replace Dummy Links in Header
Two links in 'header.php' were placeholders. I have made the following updates to them:

* The main ASU logo in the header now links to 'https://www.asu.edu'
* The 'Sign In' link in the global menu bar now links to 'https://my.asu.edu'

We are investigating how to have that 'Sign In' link actually log you in to WordPress (as currently happens in Drupal). This will at least allow users to Sign In to their ASU account - which is the true purpose of this link.