New Default Mode and Preview Color for Button
Since the UDS default for buttons is maroon, I set the default color of our block preview to maroon. Also updated the default mode so that buttons are created in 'preview' mode so you can see the button. This only works because we set a default label value in the ACF field group, by the way.