Add 'Extra Spacing' Feature on the Heading Block
Per request from the page building team, added the ability to add extra spacing on top of any header. It's a built-in value of 3rem (mt-6), which was decided upon by them and will be used very often, so this is a bit faster than typing that class in the 'advanced' box of every heading tag.