Add Multiple Buttons and Enhanced Highlighting
Updates to allow for multiple buttons in the hero area, and for highlighting only part of the title.

Includes a change to the button group to remove the instruction text (for better display in the repeater field) and apply the '-collapsed-target' class to the Link field so something shows up in 'collapsed' rows in the editor.