Change Technique for Additional Classes
After discovering a logic bug with the way I was adding the various additional classes that can be applied to a heading, I decided to overhaul the method I was using.

Rather than making a string first, and adding to it as other classes are requested, we are now adding any requested classes to an array. At the end, if there's anything in that array, we use `implode` to convert those items to a single class string. Cleaner, and doesn't rely on as many conditionals.