Allow for Missing Single Highlight Word
The original `if` statement had no `else` clause, and if the user entered a word that was not in the title, the title would not render at all! Updated this to draw a plain title in the case of the user attempting to highlight one word when that word is not in the title.