Update Hero Subtitle Markup
The sub-title text was not aligned with the other items in the hero area - most likely because it was in a column with no parent row. Added a row around the text column to offset the padding on the column, and adjusted the column sizing classes so that the sub-title area was a bit larger - to prevent unnecessary wrapping of lines.