Updated Bootstrap Markup
Added an inner `container` to the hero area to hold all the actual content (title, text, buttons), and put everything in proper rows and columns. This seems to have solved the issue with the rows only being as wide as their content, allowing the usual column sizing classes to actually work now.

I put all the content in `col-8`, to keep it as close to the 60% max width for hero content. Tested it with long titles and text, as well as short ones, and it appears to behave as expected.