Adjust Margin on Title w/No Buttons
Fixes an issue where a hero with title text, but no button, would have the title text moved down right up against the bottom of the hero image. This was because the hero content container had a negative margin on mobile to accommodate buttons.

Updated the code to apply a 'has-buttons' class when a button is present, and then targeted that class for the negative margin. When that class is absent, then the title is NOT moved down - and is, in fact, given a positive margin to keep it the same distance from the bottom as it is on other screen sizes.