Update Close Button Javascript
Updated the javascript used to close an alert. The prior javascript was broken when the markup was updated to add a FontAwesome icon. Rather than target something super-specific, like 'parentNode.parentNode.parentNode', I changed it to look for the nearest parent element of class 'alert', which should be only the top-level alert dialog.