Adjust Margins on Mobile
The existing theme was adding/adjusting margins at the smallest screen size - most likely to maximize the area available to the text - but it left the icon, text, and close button all out of alignment with each other.

For now, removing those mobile margins so that things at least don't jump around on mobile screens.