Fix Banner Widget Logic
Since our banner widget allows for the user to put the widget in the widget area - but turn off the display of the widget. When that was the case, the surrounding markup was still renderedf (even though the widget was not), leaving an unnecessary empty <div> at the top of the content area. Worse, that <div> had some margin applied, and would move content down the page.

Updated the logic in the template to only render markup when the widget is both active AND the user has specifically checked the box to display the widget.