Remove Markup From Template Partial
Because we cannot get the widget ID in the context of a basic template include, we could not query the value of the ACF field for showing/hiding the banner. Moved the surrounding markup from this template into the actual Banner widget class, where we do have access to the ACF field - and are already using it.

This prevents the surrounding markup from being rendered in the case where the widget has been dragged to the widget area, but the user has chosen to hide the widget. Before, that would only hide the widget markup, leaving unwanted wrapping markup in the final page.