Update 'Additional Classes' Code
Updated all blocks that were causing PHP warnings about undeclared variables. The source was the small piece of code in each block that retrieves the values set in the 'Additional CSS Classes' field in Gutenberg. In multiple cases, I had forgotten to declare that variable before attempting to assign a value to it.