Fix Button Highlighting Issue / Code Cleanup
While our custom formatting buttons were working, they would not adopt the 'active' state when you selected text that was already highlighted. It turns out that this feature is designed to work with SVG icons, and not the Dashicons.

I created an SVG react element from the same Dashicon icon in the code, and used that instead. I also cleaned up the indentation of the code a bit while I was there, and added some comments.