Fix Potential Issue with Menu ID
Using the 'menu' variable as the ID will not work if the menu name has any spaces in it! Updated the code to use WordPress's `sanitize_title()` method to 'slugify' the menu name to turn it into an acceptable ID.

Example: 'My sidebar menu' will become 'my-sidebar-menu' after we've made it into a slug.