Use ACF Enqueue to Load JS
As we discussed in our dev meeting on Friday, ACF has a method for only enqueueing block javascript when a block is loaded. I moved the inline script that was part of the block template into a JS file and added a line to the register.php file to enqueue it only when there's an accordion on the page.