Remove 'None' Option for Caption
Zainab suggested that we didn't need 'none' as an option, since the prime use for this block would be images with captions, and a custom or media library caption left blank would produce the same result.

Removed the 'none' option, and set the default value to  creating a custom caption.