Use @2x Version by Default
The lockup image was being slightly upscaled because it was a little bit smaller than the space in which is lives - resulting in some blurry text; especially on my Retina iMac.

For the best look on high-res screens (which is probably most screens), I updated the footer to use the 2x version by default. It's not very big, and looks a lot sharper.