Crop Innovation Footer Lockup Image
Cropped horizontal Innovation 'lockup' image so there is no space on either side of the graphic itself. This brings things into alignment on mobile, and aligns the image on desktop to the edge of the content area. That doesn't always mean it aligns with other items on the right side of the page (social icons, etc.), but it does make it fully right-aligned on the page.