Remove Conditional Logic in ACF Group
Removed the conditional logic that was hiding buttons/tags/links when the user had selected 'Yes' for hover cards. With that hover card field now deleted, this logic was not going to work (and, in fact, had reset itself to look for a value in the card title as a default behavior).