Update Logic to Avoid Single Site Issue
My code was throwing an error when used on a single site installation. I updated the logic that controls both displaying the controls, and actually switching contexts, to use WordPress's `is_multisite()` function to be sure we are on a multisite before any of this code is executed.


Also added a wrapper around the restore_current_blog() call to make sure we have actually switched blogs before switching back.