Add Support for Alternate Home URL
Added a customizer control for entering an alternate URL for the Home icon in the main nav menu. Updated the render partial that renders the main nav to look for the alternate value before rendering the home icon.

Also: removed a callback reference that was throwing an error/warning in the Query Monitor console.