Fix Editor Issue
After attempting to leave out the icon span when there was no icon, I ended up with a block that would fail in the editor if you ever deleted the icon text! I think it was because the markup was altered too much. Ended up just changing the 'mr-2' to 'mr-0' when there was no icon.