Fix Console Notice
We were getting a console notice in the Gutenberg editor - and on any page that did not have tabs - about $('nav.nav-tabs')[0] being undefined. Because the tabs wrapper and individual tabs are rendered differently in the editor, there is no first child (aka '[0]') of the tabs container. Fixed this by first checking to ensure that the container (an array) has a length greater than 0 before attempting to access the child element.

The removed the error in the console and does not appear to affect the tabs behavior.