Fix Issue With Gutenberg Block Inserter
By capturing all clicks on 'document' and sending them to the setButtonsCompatibility() method, we ended up preventing the Gutenberg block inserter from behaving correctly. The 'Browse All' option on the modal block inserter was failing to open the full block inserter as it should.

By limiting the scope of the clicks we capture to those on an actual tabbed panel - where we are trying to set focus on the selected tab - we were able to restore the block selector's desired functionality.