Require full FontAwesome icon name
Updated the block to require the full name of the FontAwesome icon per review notes. To prevent legacy icon cards from failing, we check to see if the new format is being used and use the old technique when needed.

Also added a simple bit of help text to the field making it clear what the correct format is.