Fix custom 404 page content issue
Our updated code, which uses WP_Query(), was defaulting to selecting only Posts when looking for custom 404 content - but we store that content in Pages on several sites. I updated the code to find the Page ID of the custom page, and then modified the query to use that ID to select content from a Page instead of a Post.