Move Icon Name Processing Inside Conditional
When the card header type is "icon," we have some code to handle a legacy edge case where the user was instructed to only provide part of an icon name ("user" instead of "fas fa-user" for example). A couple of deprecation notices in Query Monitor output made me realize we were processing that icon name on EVERY card on a page - even when it was not an icon card. Moved that processing code inside the existing conditional check for icon cards so that we only do this work when a card has an icon header.