Close "title and text" section by default
Updated the Card field group to not have the "title and text" section open when the card is selected. While this makes an extra step when creating a new card, it also makes it easier to get to other sections when editing a card.

It looks like ACF also added some new default values for new features.